通过豆瓣自动搜索书籍并添加

+++ src/main/java/com/mark/backend/service/IBookService.java
@@ -43,5 +43,5 @@ public interface IBookService {
 	 * @param book
 	 * @return
 	 */
-	Integer saveBook(Book book);
+	Long saveBook(Book book);
 }
